Cleaned up and rearranged code and data storage.
clean_optionmetrics.py will become a program that will parse the
zipped data into csv files.  Towards this aim, part of CallsPuts.py
has been inserted into it.  The remaining part of CallsPuts.py is
a standalone program that uses parsed data to create a graph.
This will ultiamately be a part of a single program that makes all
of the graphs.